DPR2-69 refactor - make ProductDefinitionRepository return the deserialized List<ProductDefinition> from the json file
