+++ src/main/resources/productDefinition.json
